Unify terminology on "strategy"
